pyproject.toml
[project]
name = "identity-service"
version = "0.1.0"
description = "Identity and authentication service built with FastAPI, applying Clean Architecture and the Repository pattern to support multiple persistence backends (PostgreSQL and MongoDB). Designed for scalability, testability, and real-world production scenarios."
authors = [
    {name = "Hector Lozada",email = "hlozadaccs@outlook.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi (>=0.127.0,<0.128.0)",
    "uvicorn[standard] (>=0.40.0,<0.41.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)"
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
